feat(highlight): allow disabling syntax highlighting of caller code
This change allows users to use the `JUSTEST_DISABLE_SOURCE_HIGHLIGHT`
environment variable to disable or enable the source code syntax
highlighting when showing it in failure stack traces.

Syntax highlighting remains enabled by default, but disabling it can be
useful for some testing environments without support for rendering
colors.

The environment variable can be set to the following values:
- "true", "TRUE", "True", "T", "t", "1"
- "false", "FALSE", "False", "F", "f", "0"