fix(stacktrace): avoid duplicates in stacktrace for anonymous functions
This change avoids printing the same line twice in caller stack-traces
when the direct caller to `t.Fatalf` is the same as the location that
invoked the assertion (`OrFail`, `For` or `Within`).

Such a situation can happen when the direct caller is an anonymous
function in the same line as the call to the assertion, for example:

    With(t).Verify(...).Will(MatcherFunc(func(t T) { t.Fatalf(...) })).OrFail()

In this example both the call to `OrFail` as well as the direct call to
`t.Fatalf` is on the same line, and the caller stack trace would print
this line twice. This change prevents that.