feat: support slowing down tests to enable debugging consequences
This change introduces a supported environment variable called
`JUSTEST_SLOW_FACTOR` which, when set to an integer value, will cause
Justest to multiple durations passed to the `For(...)` and `Within(...)`
methods with its value. This will cause these duration to be longer,
thus allowing the developer to investigate or debug side consequences of
such tests - like logging into clusters and checking the effects of
end-to-end tests, inspecting file-systems, etc.