refactor: refactor fluent API & source printing for better readability
This change refactors the fluent API of the package to make the test
source code flow better in terms of readability. For example, instead
of:

    With(t).Verify(1).Will(EqualTo(2)).OrFail()

The API is now either:

    With(t).EnsureThat("1 equals 2").ByVerifying(1).Will(EqualTo(2)).Now()
    With(t).VerifyThat(1).Will(EqualTo(2)).Now()

The call to "EnsureThat" is optional and is meant to help documenting
the test in case of failures. The change from ".OrFail()" to ".Now()"
ensures that every assertion ends with a timing specification:

- With(t).... .Now()
- With(t).... .For(...)
- With(t).... .Within(...)

Additionally, this refactor improves the printing of the assertion's
source code location, enabling support for multi-line location source
code. For instance, for the following assertion:

    With(t).
        EnsureThat("...").
        ByVerifying(1).
        Will(EqualTo(2)).
        Now()

previous versions only printed the line containing the "Now()" call, but
this version will print the entire statement (from the "With(t)" up to
and including the call to "Now()".)